+++ apps/vmq_server/src/vmq_queue.erl
@@ -222,7 +222,7 @@ online({add_session, SessionPid, #{allow_multiple_sessions := false} = Opts}, Fr
     %% and wait with the reply until all the sessions
     %% have been disconnected
     #state{id = SubscriberId} = State,
-    lager:warning("SubscriberId ~p disconnected by same client_id", [SubscriberId]),
+    lager:info("client ~p disconnected due to multiple sessions not allowed", [SubscriberId]),
     disconnect_sessions(State),
     {next_state, state_change(add_session, online, wait_for_offline),
      State#state{waiting_call={add_session, SessionPid, Opts, From}}};