+++ apps/vmq_server/src/vmq_server_cli.erl
@@ -59,6 +59,7 @@ register_cli() ->
     vmq_cluster_leave_cmd(),
     vmq_cluster_upgrade_cmd(),
 
+    vmq_mgmt_add_api_key_cmd(),
     vmq_mgmt_create_api_key_cmd(),
     vmq_mgmt_delete_api_key_cmd(),
     vmq_mgmt_list_api_keys_cmd(),
@@ -86,6 +87,7 @@ register_cli_usage() ->
 
     clique:register_usage(["vmq-admin", "api-key"], api_usage()),
     clique:register_usage(["vmq-admin", "api-key", "delete"], api_delete_key_usage()),
+    clique:register_usage(["vmq-admin", "api-key", "add"], api_add_key_usage()),
     ok.
 
 vmq_server_stop_cmd() ->
@@ -385,6 +387,21 @@ vmq_mgmt_create_api_key_cmd() ->
                end,
     clique:register_command(Cmd, KeySpecs, FlagSpecs, Callback).
 
+vmq_mgmt_add_api_key_cmd() ->
+    Cmd = ["vmq-admin", "api-key", "add"],
+    KeySpecs = [{'key', [{typecast, fun(Key) ->
+                                            list_to_binary(Key)
+                                    end}]}],
+    FlagSpecs = [],
+    Callback = fun(_, [{'key', Key}], _) ->
+                       vmq_http_mgmt_api:add_api_key(Key),
+                       [clique_status:text("Done")];
+                  (_,_,_) ->
+                       Text = clique_status:text(api_add_key_usage()),
+                       [clique_status:alert([Text])]
+               end,
+    clique:register_command(Cmd, KeySpecs, FlagSpecs, Callback).
+
 vmq_mgmt_delete_api_key_cmd() ->
     Cmd = ["vmq-admin", "api-key", "delete"],
     KeySpecs = [{'key', [{typecast, fun(Key) ->
@@ -393,7 +410,10 @@ vmq_mgmt_delete_api_key_cmd() ->
     FlagSpecs = [],
     Callback = fun(_, [{'key', Key}], _) ->
                        vmq_http_mgmt_api:delete_api_key(Key),
-                       [clique_status:text("Done")]
+                       [clique_status:text("Done")];
+                  (_,_,_) ->
+                       Text = clique_status:text(api_delete_key_usage()),
+                       [clique_status:alert([Text])]
                end,
     clique:register_command(Cmd, KeySpecs, FlagSpecs, Callback).
 
@@ -521,9 +541,10 @@ metrics_show_usage() ->
 
 api_usage() ->
     ["vmq-admin api-key <sub-command>\n\n",
-     "  Create, delete, and show API keys for the HTTP management interface.\n\n",
+     "  Create, add, delete, and show API keys for the HTTP management interface.\n\n",
      "  Sub-commands:\n",
      "    create      Creates a new API key.\n",
+     "    add         Adds a new API key.\n",
      "    delete      Deletes an existing API key.\n",
      "    show        Shows all API keys.\n"
     ].
@@ -531,6 +552,11 @@ api_usage() ->
 api_delete_key_usage() ->
     ["vmq-admin api-key delete key=<API Key>\n\n",
      "  Deletes an existing API Key.\n\n"
+    ].
+
+api_add_key_usage() ->
+    ["vmq-admin api-key add key=<API Key>\n\n",
+     "  Adds an API Key.\n\n"
     ].
 
 ensure_all_stopped(App)  ->