+++ apps/vmq_server/test/vmq_publish_SUITE.erl
@@ -421,8 +421,8 @@ not_allowed_publish_close_qos2_mqtt_3_1_1(_) ->
     {error, closed} = gen_tcp:recv(Socket, 0, 1000).
 
 message_size_exceeded_close(_) ->
-    OldLimit = vmq_config:get_env(message_size_limit),
-    vmq_config:set_env(message_size_limit, 1024, false),
+    OldLimit = vmq_config:get_env(max_message_size),
+    vmq_config:set_env(max_message_size, 1024, false),
     Connect = packet:gen_connect("pub-excessive-test", [{keepalive, 60}]),
     Connack = packet:gen_connack(0),
     Publish = packet:gen_publish("pub/excessive/test", 0, vmq_test_utils:rand_bytes(1024),
@@ -432,7 +432,7 @@ message_size_exceeded_close(_) ->
     ok = gen_tcp:send(Socket, Publish),
     {error, closed} = gen_tcp:recv(Socket, 0, 1000),
     true = lists:member({counter, mqtt_invalid_msg_size_error, 1}, vmq_metrics:metrics()),
-    vmq_config:set_env(message_size_limit, OldLimit, false).
+    vmq_config:set_env(max_message_size, OldLimit, false).
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %%% Hooks