+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Log when a client is disconnected and multiple sessions are not allowed.
 - Fix tracer `mountpoint` parameter bug.
+- Make it possible to add/inject HTTP API keys.
 
 ## VERNEMQ 1.1.1